+++ avutil/interfaces.go
@@ -1,0 +1,6 @@
+package avutil
+
+type IDictionary interface {
+	Set(key, value string) error
+	Free()
+}